Improve jacobian calculation

+++ src/panel.jl
@@ -113,7 +113,7 @@ function init_pos!(
     panel.x_airf .= x_airf
     panel.y_airf .= y_airf
     panel.z_airf .= z_airf
-    panel.delta = Float64(delta)
+    panel.delta = delta
     return nothing
 end
 
@@ -361,7 +361,7 @@ function calculate_cd_cm(panel::Panel, alpha::Float64)
     elseif panel.aero_model == POLAR_VECTORS
         cd = panel.cd_interp(alpha)::Float64
         cm = panel.cm_interp(alpha)::Float64
-    elseif panel.aero_model == POLAR_MATRICES    
+    elseif panel.aero_model == POLAR_MATRICES
         cd = panel.cd_interp(alpha, panel.delta)::Float64
         cm = panel.cm_interp(alpha, panel.delta)::Float64
     elseif !(panel.aero_model == INVISCID)
